+++ src/seppl/__init__.py
@@ -1,6 +1,7 @@
 from ._plugin import Plugin, PluginWithLogging, OutputProducer, InputConsumer, check_compatibility, LoggingHandler, Initializable, init_initializable
 from ._args import split_cmdline, split_args, args_to_objects, is_help_requested, enumerate_plugins, escape_args, unescape_args
 from ._registry import Registry, MODE_DYNAMIC, MODE_EXPLICIT, MODES
+from ._class_registry import ClassListerRegistry, get_class_lister, DEFAULT
 from ._entry_points import generate_entry_points
 from ._help import generate_plugin_usage, generate_help, HELP_FORMATS, HELP_FORMAT_TEXT, HELP_FORMAT_MARKDOWN
 from ._types import get_class, get_class_name, classes_to_str, AnyData